requirements.txt
langgraph==0.5.2
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
playwright==1.53.0
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
starlette==0.47.1
tenacity==9.1.2